Contacts Service

+++ backoffice/src/main/java/com/codemovers/scholar/v1/backoffice/db/entities/Contacts.java
@@ -44,12 +44,12 @@ public class Contacts implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Long id;
+    private Integer id;
     @Size(max = 255)
     @Column(name = "parent_type")
     private String parentType;
     @Column(name = "parent_id")
-    private BigInteger parentId;
+    private Integer parentId;
     @Size(max = 255)
     @Column(name = "contact_type")
     private String contactType;
@@ -63,15 +63,15 @@ public class Contacts implements Serializable {
     public Contacts() {
     }
 
-    public Contacts(Long id) {
+    public Contacts(Integer id) {
         this.id = id;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
@@ -83,11 +83,11 @@ public class Contacts implements Serializable {
         this.parentType = parentType;
     }
 
-    public BigInteger getParentId() {
+    public Integer getParentId() {
         return parentId;
     }
 
-    public void setParentId(BigInteger parentId) {
+    public void setParentId(Integer parentId) {
         this.parentId = parentId;
     }
